+++ test/trace/tracestate-validators.test.ts
@@ -15,7 +15,7 @@
  */
 
 import * as assert from 'assert';
-import { validateKey, validateValue } from '../../src/trace/internal/validators';
+import { validateKey, validateValue } from '../../src/trace/internal/tracestate-validators';
 
 describe('validators', () => {
   describe('validateKey', () => {
